Moved the Builder class to its own file.
