+++ tools/clang/lib/AST/CMakeLists.txt
@@ -36,7 +36,8 @@ add_clang_library(clangAST
   ExprConstant.cpp
   ExprCXX.cpp
   ExternalASTSource.cpp
-  HlslTypes.cpp
+  HlslBuiltinTypeDeclBuilder.cpp # HLSL Change
+  HlslTypes.cpp # HLSL Change
   InheritViz.cpp
   ItaniumCXXABI.cpp
   ItaniumMangle.cpp
